adjusted romloader, fixed error in cpu

+++ src/test/java/group1/mips_simulator/components/instruction/InstructionTest.java
@@ -25,5 +25,5 @@ class InstructionTest {
         assertEquals((short) 0, hlt.toShort());
 
         Instruction ldr = factory.buildInstruction_fromBinary("0000010101101101");
-        assertEquals((short) 0b0000010101101101, ldr.toShort());    }
+        assertEquals( 0b0000010101101101, ldr.toShort());    }
 }
